+++ packages/f-footer/CHANGELOG.md
@@ -3,6 +3,18 @@
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+v4.1.0
+------------------------------
+*September 28, 2020*
+
+### Added 
+- data-test IDs to Footer Component 
+- Test for Footer Icons 
+- F-Footer Component-Object 
+
+### Changed
+- Npm scripts 
+
 
 v4.0.0
 ------------------------------